Task3 feature testing

+++ app/controllers/questions_controller.rb
@@ -1,18 +1,24 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :question, only: %i[show edit update destroy]
+  before_action :authenticate_user!, except: %i[index show]
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answer = Answer.new
+  end
 
-  def new; end
+  def new
+    @question = Question.new
+  end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.build(question_params)
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Question created successfully'
     else
       render :new
     end
@@ -29,17 +35,19 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'question deleted successfully'
+    else
+      redirect_to questions_path, alert: 'it is not your question'
+    end
   end
 
   private
 
   def question
-    @question ||= params[:id] ? Question.find(params[:id]) : Question.new
+    @question = Question.find(params[:id])
   end
-
-  helper_method :question
 
   def question_params
     params.require(:question).permit(:title, :body)
